fix: style 파일 통일
