+++ src/views/ApplyPage/components/BottomSection/index.tsx
@@ -1,0 +1,5 @@
+const BottomSection = () => {
+  return <div>BottomSection</div>;
+};
+
+export default BottomSection;
