+++ src/views/ApplyPage/components/PartSection/index.tsx
@@ -1,9 +1,9 @@
 import { UseFormReturn } from 'react-hook-form';
 
 import Textarea from '@components/Textarea';
+import { sectionContainer, title } from 'views/ApplyPage/style.css';
 
 import { PART_QUESTION } from './constants';
-import { container, title } from './style.css';
 
 const PartSection = ({ formObject }: { formObject: Pick<UseFormReturn, 'register' | 'formState' | 'watch'> }) => {
   const {
@@ -12,7 +12,7 @@ const PartSection = ({ formObject }: { formObject: Pick<UseFormReturn, 'register
     formState: { errors },
   } = formObject;
   return (
-    <section className={container}>
+    <section className={sectionContainer}>
       <h2 className={title}>파트별 질문</h2>
       {PART_QUESTION.map(({ question, maxCount, required }, id) => (
         <div key={question}>
